@@ -1,9 +1,13 @@
 package de.l3s.icrawl.api.reddit;
 
-import org.joda.time.DateTime;
+import java.time.Instant;
+import java.time.OffsetDateTime;
+
 import org.junit.Test;
 
-import de.l3s.icrawl.api.reddit.RedditDateTimeDeserializer;
+import de.l3s.icrawl.api.reddit.RedditDateTimeModule.RedditInstantDeserializer;
+
+import static java.time.ZoneOffset.UTC;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.CoreMatchers.isA;
 import static org.junit.Assert.assertThat;
@@ -13,15 +17,12 @@ public class RedditDateTimeDeserializerTest {
     @Test
     public void testParseTimestamp() {
         // equal to 2014-08-01T04:46:24-07:00, is off by an hour
-        float ts = 1406897184.0f;
-        RedditDateTimeDeserializer deserializer = new RedditDateTimeDeserializer();
+        double ts = 1406897184.0;
 
-        DateTime parsed = deserializer.parseTimestamp(ts);
+        Instant parsed = RedditInstantDeserializer.parseTimestamp(ts);
         System.out.println(parsed);
-        assertThat(parsed, isA(DateTime.class));
-        assertThat(parsed.getYear(), is(2014));
-        assertThat(parsed.getMonthOfYear(), is(8));
-        assertThat(parsed.getDayOfMonth(), is(1));
+        assertThat(parsed, isA(Instant.class));
+        assertThat(parsed, is(OffsetDateTime.of(2014, 8, 1, 12, 46, 24, 0, UTC).toInstant()));
     }
 
 }
